Add a new in-memory cache for TreeAuxData in ObjectStore
Summary:
Add a new in-memory cache for TreeAuxData in ObjectStore. This is the Tree version of the auxDataCache.
It will be used in the following diff.

## Context
Currently we fetch TreeAuxData directly from the backing store without caching. We want to add cache support. The cache will be consist of the in-memory cache (in ObjectStore) as well as the local store.

Reviewed By: jdelliot

Differential Revision: D72261663

fbshipit-source-id: 7ceab55eefbcef6957e2ad88273d840bd829b562