Implement methods to fetch TreeAuxData from in-memory cache
Summary:
Implement the method to fetch TreeAuxData from in-memory cache.
This method will be called by `getTreeAuxData()`.

## Context
Currently we fetch TreeAuxData directly from the backing store without caching. We want to add cache support. The cache will be consist of the in-memory cache (in ObjectStore) as well as the local store.

Reviewed By: jdelliot

Differential Revision: D72263450

fbshipit-source-id: e2f5dc7d53f35369090e790d1051135b47c704ab