Move AI split button next to title, emphasize on first commit
Summary:
I played with this a little and found moving this button to the title bar fits a little nicer, using less vertical space.

If we use a more compact label for the button, this fits in even nicer.

At the same time, we can change the emphasis on this button when there's exactly one commit.

Lastly, we add tooltips to make it a bit more approachable.

Reviewed By: quark-zju

Differential Revision: D72427656

fbshipit-source-id: fa83cb197c6d8c7e636938ae6fe689dc9249bdb8