store: cache trees post case sensitivity
Summary:
This diff changes when we call `changeCaseSensitivity()` while interacting with the `TreeCache`.

Before - Cache Hit:
* return changeCaseSensitivity(tree)

After - Cache Hit:
* return changeCaseSensitivity(tree)

Before - Cache Miss (write fetched tree to TreeCache):
* insert tree
* changedTree = changeCaseSensitivity(tree)
* return changedTree

After - Cache Miss (write fetched tree to TreeCache):
* changedTree = changeCaseSensitivity(tree)
* insert changedTree
* return changedTree

From xavierd describes the perf reason for this change:
> Ideally Eden inserts into the cache once, but reads from it many times. Since changing the case sensitivity is currently done after the read, the number of times the case sensitivity is changed is O(n), while if we move it before inserting, it becomes O(1)

As pointed out in discussion on this diff - if two mounts for the same repo, with different case sensitivities, fall into a pattern of:
* cache miss A
* cache read B
* cache miss B
* cache hit A

We will still see the O(n) behavior ( `changeCaseSensitivity()` is a no-op if the sensitivity doesn't need to be changed).

Potential followups needed:
* two tree caches or a tree cache per mount to completely avoid `changeCaseSensitivity()` on read altogether with the tradeoff of increased memory usage. Or some different key that includes the sensitivity.
* verifying we correctly handle case sensitivity at the local store layer

Reviewed By: quark-zju

Differential Revision: D71851987

fbshipit-source-id: 15fc450a0316751bedac36fd9e76be91c6fd5541