Error logging improvements in the edenapi sender
Summary:
The goal of the diff is to ensure that when we return an error 1) we have enough error to debug and 2) it's definitely a protocol error and not a programming error or similar. In practical terms, this means we can confidently retry (if that's enabled).

A couple of things going into this:

* silently ignore errors when trying to obtain stats;
* add a context everywhere we can legitimately return an error.

Reviewed By: lmvasquezg

Differential Revision: D72401107

fbshipit-source-id: 2c1eb3036479f6b57874a0c9a61f25f5fe343288