Remove megarepolib dependency from cross_repo_sync_test_utils
Summary:
# This stack
**Please see D70781050 before reviewing this to get context on the plan and the goal.**

## This diff
Makes some changes to the `cross_repo_sync_test_utils` crate, to unblock D71973796, which moves it into `cross_repo_sync`.

This diff is needed to remove a dependency to `megarepolib`, which creates a cyclic dependency in Cargo.

But the change itself is reasonable, because the current megarepo merge process doesn't do what `init_small_large_repo` is doing now.

`init_small_large_repo` is currently creating a noop config to first import the commits. It then creates a **MOVE** commit, to move all the paths to the new path. After that it sets working copy equivalence and changes the mapping. This was probably the old proces.

The current process already sets the prefix in the noop config, removing the need for a move commit.

Differential Revision: D72450471

fbshipit-source-id: ca2b1e7b596ba8a2cef9ab7d18e7bb75355deffc