Fix for socket exhaustion preventing eden test startup: Retry setup_eden_test
Summary:
# Context
S506089
Lego-windows has been deprecated, and in order to get edenfs-windows-release working on it's replacement EWW, we need to resolve the tests failing on it.

Specifically, the tests are failing due to the test environment experiencing socket exhaustion, which causes eden to fail to start properly.
https://learn.microsoft.com/en-us/troubleshoot/windows-client/networking/tcp-ip-port-exhaustion-troubleshooting

# This Diff
This diff introduces retries for setup_eden_test. This is part of a series of retries required to properly get eden to start correctly. This cannot be covered by the --retries flag inside the test agent because those resume immediately, without giving enough time for sockets to become available. This is being deployed in conjunction with a config change that reduces the time it takes for a socket to become freed after closing from 4m to 30s https://www.internalfb.com/diff/D72474652

# Why this function
When ephemeral sockets are exhausted, eden will fail to start for various reasons. This retry covers the following cases:
 - When running clone, due to being unable to create a ProjFSChannel (P1773665417). This is after eden.start succeeds and kills the eden daemon instance, requiring a restart, but the existing state makes it difficult to resume from a lower level

# Detail

# Changes
 - When setup_eden_test fails, retry it a certain number of times
 -- shutdown the eden instance, and then retry from the beginning inside a new tmp folder.
 - Each new test is started in a new tmp folder to prevent read/write issues. tmp_dir has been updated to reflect this
 - replace mkdir with makedirs(exist_ok=True) for similar reasons

Differential Revision: D72179207

fbshipit-source-id: 39c34350620fc7fe2ad600ff7509af85e1f39833