Split of "[Eden] Fix for socket exhaustion preventing eden test startup" : Handle errors during shutdown
Summary:
# Context
S506089
Lego-windows has been deprecated, and in order to get edenfs-windows-release working on it's replacement EWW, we need to resolve the tests failing on it.

Specifically, the tests are failing due to the test environment experiencing socket exhaustion, which causes eden to fail to start properly.
https://learn.microsoft.com/en-us/troubleshoot/windows-client/networking/tcp-ip-port-exhaustion-troubleshooting

This test is sufficient alongside D72474652 to pass the worker, but there are a couple of other issues that get addressed in followup diffs that are getting masked by the --retry flag in the release build process

# Detail

Not being addressed at this current level
 - Calling eden start outside of setup_eden_test can fail
 - Extra long timeouts caused by run_cmd

# This Diff
Despite using eden.kill in the previous function, which checks if the process is alive, it's still possible to sometimes get a 'edenfs is not running' error. This is due to several possibilites.
Fixes these possible cases when running eden shutdown with retries enabled
 - 'edenfs is not running': Occurs when shutdown is called after eden has crashed. The eden daemon is not responding, but the process is still running. (eg. https://www.internalfb.com/intern/everpaste/?handle=GN9gQRbo2msF4D4GAP4gG4gehaxsbswMAAAz.
- 'WinError 10054' - Cause of the previous. Caused when trying to create a PrjFSChannel when out of sockets, and will kill the eden process.
 - WinError 10055: No sockets are available when the CLI attempts to create a socket to communicate with eden. Formerly fixed here but rare and gets addressed in via the previous diff

# Detail

# Changes
 - Copy report_time from testcase.py to edenclient.py, to improve time tracking.
 - Add retry flag to shutdown/kill.
 - When retry flag is set, use a much lower timeout to wait for eden to shut down in response to eden stop
 - Add try-catch to running "eden stop" inside shutdown. Catch the error 'edenfs is not running', and if retry flag is set, do not fail due to wait timeout.

Reviewed By: quark-zju

Differential Revision: D72418860

fbshipit-source-id: a7ea690e3bd75655d340410d9f06102468b9e61a