Log total trees and files served to scuba
Summary:
In

https://www.internalfb.com/code/configerator/source/scm/mononoke/ratelimiting/ratelimits.cconf?lines=18-35
We define a ratelimit that should actually be a load shedding limit. In order to make the switch I need the data to be in ODS

Once the data gets published to ODS I can set up the equivalent load shedding limit and remove this one

Note that Commits are currently only logged on the Wireproto path. I want to understand usage there before I set up ODS counters as I've seen some counters not get published at all.

Reviewed By: RajivTS

Differential Revision: D72556506

fbshipit-source-id: 96d334f304e93eaca75e557c8dc40e0cb488becc