Delete most `runtime_deps` in the repo
Summary:
`python_*` rules accept a parameter called `runtime_deps`, in case you don't know yet, take a moment now to make a prediction as to what adding a target to this dependency list does. Got it?

Having any `runtime_deps` causes the macro to generate a `cpp_library` target, depend on that target from the python target, and then add the runtime deps as a dependency to that cpp library. Not only do I have no idea how that makes something a runtime dep, but it also has the wonderful property that runtime deps which don't output providers recognized by a cpp library are completely ignored.

People usually passed binaries into here, so that turns out to be almost all of them. These can all be deleted, and that's what this diff does.

I'm pretty confident that this is a nop since I only touched cases where this pointed to targets that would get ignored.

There are a couple cases left besides that, I'll deal with those after

#buildmore

Reviewed By: zsol

Differential Revision: D72542581

fbshipit-source-id: ee8a11d8b1d5140a252a9dccccdc9af25a5d20aa