Extract types for diffSplit to a separate file
Summary:
The diffSplit implementation in isl-server was importing from isl/src/stackEdit/diffSplit. This also imported other stuff, which confused typescript.

At actual bundle time this is a non-issue as it's a type import. But `tsc --noEmit` thought it had to traverse the file and typecheck several `isl/` files using the isl-server tsconfig, which is invalid (e.g. tsx files all fail.

So this diff fixes the CI job but there shouldn't be anything that was broken at runtime.

Reviewed By: zzl0

Differential Revision: D72603407

fbshipit-source-id: a1ba59cfd07aa21062a97d8a88d5f0d4153e8b98