Record clicked & accepted ai splits
Summary:
Record metrics for each time you click the AI suggest button, and also count at the very end how many `splitWithAI` edit OPs appear in the history.

The goal is to understand acceptance rate of AI Splits.

Acceptance rate = (# splits in history, where the user clicks save) / (# clicks of AI Split, regardless of if they save or cancel)

Because you can undo / redo, it's insufficient to bump metrics when you get an AI result as "success". (I think this actually also makes some of the other metrics inaccurate, but for those it's still useful to know if you used it, even if you undo later)

Reviewed By: quark-zju

Differential Revision: D72609210

fbshipit-source-id: 6f07c750a9ceba1f77ceab3353fd7b8d35446a0a