TakeoverClient: add simulated errors for testing
Summary:
# This diff

Adds a new config that will force the TakeoverClient to crash prior to receiving takeover data from the existing server.

# Context

Crashes during graceful restart are plaguing macOS. Although these happen in the background, these greatly affect users because they cause Eden daemons to exit uncleanly. Because of the unclean exit, mount points get left in the mount table and cause all sorts of issues. The biggest issues are:

1) Any attempts to perform IO into the mount point fail
2) Any subsequent attempts to remount the mount point fail. These attempts fail because of #1, and Eden's startup/mount codepaths haven't been taught how to deal with this.

The goal of this stack is to:

1) Introduce a mechanism to easily repro graceful restart crashes (both crashes that occur in the TakeoverClient and TakeoverServer)
2) Use the new mechanism to troubleshoot issue #2 above
3) Use the learnings from #2 to make EdenFS automatically recover from graceful restart crashes (unmount stale mounts and successfully remount any mounts that were mounted prior to the crash).

Reviewed By: jdelliot

Differential Revision: D72490700

fbshipit-source-id: f86b7d1317ce8b5accbca3c6102d83e277b213ce