Move shared type and helper function to appropriate mods within sapling-client
Summary:
# Context:

We are working to refactor the EdenFs CLI client library into a supported XFN consumable library to avoid code duplication and sub-optimal usage patterns.

# This Diff:

In subsequent diffs we will be adding support `sl debugdiffdirs`. Here we are moving related shared types and functions to appropriate mods in preparation for that.

Reviewed By: MichaelCuevas

Differential Revision: D72661507

fbshipit-source-id: 36d771224413adb133022ad855644b17c23d343a