Change DB used for checkpointing table
Summary:
The `repo_metadata_logger` tool relies on a form of checkpointing mechanism to ensure that we make incremental progress over time. The checkpointing state is stored in a XDB table and we write very frequently to it. To avoid this problem, we decided to move this table to a different DB thereby bypassing the write contention issue.

NOTE: Only to be landed once backfilling of the new DB gets completed

Reviewed By: clara-9

Differential Revision: D72459943

fbshipit-source-id: 2acb292de31bfd807b2b5ea5d0dc1fb9a9d7d222