Check the in-memory cache for TreeAuxData
Summary:
Actually use the cache. Check TreeAuxData from the in-memory cache first. Insert into the cache after fetching the data from the backing store.

In the following diffs we will implement the step to check the local store for Tree aux data before consulting the backing store.

## Context
Currently we fetch TreeAuxData directly from the backing store without caching. We want to add cache support. The cache will be consist of the in-memory cache (in ObjectStore) as well as the local store.

Reviewed By: MichaelCuevas

Differential Revision: D72263457

fbshipit-source-id: 6b5efc97c346814b0e6fe006d02b4b02e58a6323