Re-enable TreeAuxDataFamily
Summary:
Add a config for localStoreTreeAuxSizeLimit, currently using the same value as localStoreTreeSizeLimit. This is a preparation for the following diffs which implement the local store for Tree aux data.

## Context
Currently we fetch TreeAuxData directly from the backing store without caching. We want to add cache support. The cache will be consist of the in-memory cache (in ObjectStore) as well as the local store.

Differential Revision: D72268972

fbshipit-source-id: 0160d0f633e1266bae12eaa70a5e85c1c455fe7a