Implement data structure to hold serialized TreeAuxData
Summary:
Add data structure to hold serialized TreeAuxData. This is needed when we store the retrieved TreeAuxData into the local store.

This is mostly similar to SerializedBlobAuxData

## Context
Currently we fetch TreeAuxData directly from the backing store without caching. We want to add cache support. The cache will be consist of the in-memory cache (in ObjectStore) as well as the local store.

Reviewed By: jdelliot

Differential Revision: D72413769

fbshipit-source-id: 6d558b303401c496a580788a241f5c65217d3ec7