Add Local Store method to fetch TreeAuxData
Summary:
Add methods to put/store TreeAuxData into/from local store.
They will be used in the following diff.

## Context
Currently we fetch TreeAuxData directly from the backing store without caching. We want to add cache support. The cache will be consist of the in-memory cache (in ObjectStore) as well as the local store.

Reviewed By: jdelliot

Differential Revision: D72425813

fbshipit-source-id: f72eae43e2f1baa68bfec52fdf036f5558f7c746