Check local store before fetching Tree aux data from backing store
Summary:
When retrieving Tree aux data, check the local store first (when config allows). If we can't find it from the local store, then retrieve from the backing store. Once we have the result, cache it into the local store(when config allows).

## Context
Currently we fetch TreeAuxData directly from the backing store without caching. We want to add cache support. The cache will be consist of the in-memory cache (in ObjectStore) as well as the local store.

Reviewed By: jdelliot

Differential Revision: D72425820

fbshipit-source-id: 6d165cb347e93675a29f719c64ceca88d093e4f3