Add basic error support for sapling-client
Summary:
# Context:

We are working to refactor the EdenFs CLI client library into a supported XFN consumable library to avoid code duplication and sub-optimal usage patterns.

# This Diff:

In subsequent diffs we will need Results (and their errors) that can be Cloned. anyhow Errors don't fit the bill. Let's fix that with a very simpe and straightforward set of Errors. We can evolve and improve these over time.

Reviewed By: lXXXw

Differential Revision: D72832081

fbshipit-source-id: cdb93e7e532415bd3b021e8e559a3f6582b016d7