mutations: allow using read-committed transaction level gated behind a JK
Summary:
We've been seeing lock wait timeouts / deadlocks for some time when inserting new entries into the mutations table. Switching transaction isolation level to read-committed is one the most recommended mitigations for this and the insertion queries do not rely on behaviours of repeatable read so it's worth giving it a try.

Inspired from Youssef's Commit graph diff D68417791

Reviewed By: lmvasquezg

Differential Revision: D72854253

fbshipit-source-id: 4549a35212f79e031e4cdb2196e38b62c0321e2e