edenapi: rename "maxrequests" config to "max-concurrent-requests"
Summary:
Hyphens is our preferred config key style, and I think adding "concurrent" is more clear what it does.

I don't think it is safe to just change the name even though it is only referenced within Sapling. This is because eden also uses this code, but does not necessarily have a dynamic config in sync with its version of the sapling code.

Reviewed By: zzl0

Differential Revision: D72823374

fbshipit-source-id: 6cc1d201be7a6b1a0bb6b3315e960a74d77dcfb4