Add caching of mergebase computation
Summary:
# Context:

We are working to refactor the EdenFs CLI client library into a supported XFN consumable library to avoid code duplication and sub-optimal usage patterns.

# This Diff:

Watchman had some caching of frequently ran Sapling operations. This will be handly for clients that are more long-running (e.g. buck2 and meerkat in daemon mode). Here we are adding caching to mergebase results.

Reviewed By: MichaelCuevas

Differential Revision: D72898031

fbshipit-source-id: 35a7ec0e1028c1220f80829b9483b68ebeaf6275