set a flag for intentional unmounts
Summary:
## Background

EdenFS can become unmounted accidentally for various reasons. Many users resort to running `eden doctor` to address this issue.
The proposed solution involves implementing a periodic task within EdenFS to monitor the mount status. If an accidental unmount is detected, the system will automatically remount the repository. This approach aims to provide a seamless recovery process, reducing the need for manual intervention and improving overall user experience. To ensure that intentional user requested unmounts are respected, a flag will be introduced to distinguish between intentional and accidental unmounts. More explanations: https://fb.workplace.com/groups/1708850869939124/permalink/1930486064442269/

## This diff
In all the intentional unmount which are triggered from CLI set a flag that the unmount in intentional. The flag is set by creating a file `intentionally-unmounted` in the `client_dir`
The file should removed in the beginning of each mount to reset this flag

Reviewed By: lXXXw

Differential Revision: D72203255

fbshipit-source-id: 3cc010454cf68853028422bdc1758ed1d1c008ca