unamend: raise error if ctx parents differ
Summary:
It doesn't make sense to "unamend" after a rebase, for example. That would reset your working copy to that of the predecessor commit _before_ the rebase (which could include an arbitrary amount of unrelated changes).

Now instead we raise an error with some pointers to other useful commands.

I left a config knob in to disable this new check if needed.

Reviewed By: quark-zju

Differential Revision: D72990340

fbshipit-source-id: 510c3237a70edd6e841d25ebcbc6f99e5c1cdd22