Make the EdenAPI timeout configurable
Summary:
Long timeouts are very visible in benchmarks: a single failed upload will stall the entire pipeline for the duration of the timeout. We need to be able to detect and recover faster.

I've tested the timeout for contents upload down to 20s, and it seems to work well. I haven't tested it at all for other endpoints. Using a JK is the safe path, it lets us experiment without having to go through a release cycle. Changing the JK will only take effect at the next restart which is even safer.

Reviewed By: lmvasquezg

Differential Revision: D73041892

fbshipit-source-id: a55bbd3972d4696b827ac3b448a545170b3f3eed