commit cloud: use CloudChangesetId in snapshots field
Summary:
## This stack

Doing the needed refactorings to make the commit cloud facet changeset-type agnostic. This means moving all `HgChangesetId` usages to be only on SLAPI and keep the processing abstracted to Sha1 hashes.

## This diff

Use `CloudChangesetId` on snapshots. This works fine for me and AFAIK not that many people use it so low risk to migrate these for now.

Reviewed By: RajivTS

Differential Revision: D73039172

fbshipit-source-id: dd8d479e56da0ef18e403fb938b1665ccaf49de9