commit cloud: use CloudChangesetId in checkout locations
Summary:
## This stack

Doing the needed refactorings to make the commit cloud facet changeset-type agnostic. This means moving all `HgChangesetId` usages to be only on SLAPI and keep the processing abstracted to Sha1 hashes.

## This diff

Checkout locations is an unused intern project feature. We kept some of it in the new service but given it's not exposed there's no risk in migrating it to the new type of changeset id.

Reviewed By: RajivTS

Differential Revision: D73040025

fbshipit-source-id: 57645a647520bba2d0a251313ff0ee9a936f437f