commit cloud types: use Result rather than anyhow::Result
Summary:
## This stack

Doing the needed refactorings to make the commit cloud facet changeset-type agnostic. This means moving all `HgChangesetId` usages to be only on SLAPI and keep the processing abstracted to Sha1 hashes.

## This diff

I generally prefer having `anyhow::Result` for specifying `Result` type but this is not in line with the general mononoke guidelines, so changing accordingly

Reviewed By: RajivTS

Differential Revision: D72869887

fbshipit-source-id: d62ca5c250085db4f018bb25dcea93aee6a056e0