commit cloud: introduce cloudchangesetid
Summary:
## This stack

Doing the needed refactorings to make the commit cloud facet changeset-type agnostic. This means moving all `HgChangesetId` usages to be only on SLAPI and keep the processing abstracted to Sha1 hashes.

## This diff

Introducing the type `CloudChangesetId` which can be either an Hg hash or a Git hash and will be decoded according to repo config. I also added the `ConvIr` implementation to allow writing to/from DB with that type directly and doing the appropriate hex-encoding following the commit cloud database format.

The idea is switching from having a hex-encoded prod DB and a non-encoded test DB and have them both be hex encoded to have more accurate testing.

Reviewed By: RajivTS

Differential Revision: D72869862

fbshipit-source-id: c267f80bf4b07ae712b917e87624ffc58d15aa26