Add setting to hide irrelevant stacks
Summary:
If an entire stack consists of commits that are outside your cwd, then the whole stack is irrelevant and we can let the user hide it, if they enable a setting.

I chose to make this setting not enabled by default, as I think it's useful to see all the commits in some cases. Maybe we can add an upsell or something for this.

**Open to other opinions on if this should be set to "hide" by default.**

Note that we need some trickery in computing the DAG where we hide stacks that entirely irrelevant. There may be a better way to describe the DAG selection for "stacks that only contains commits passing this predicate"—the best I found is to populate all the descendants of each root and then filter from there.

The settings UI exposes this as a dropdown, but it's stored as two booleans, since we already have one boolean. This means we respect the existing setting without needing to do a data migration.

Reviewed By: quark-zju

Differential Revision: D72682667

fbshipit-source-id: e5676daf87a7113f73a26e57a61861203fc2e15e