factory: relax FunctionSignature::In lifetime
Summary:
Rust requires both `In` and `Out` to be `'static` with the `fn(In) -> Out:
'static` constraint (which seems overly strict). Relaxing the `In` type
to be not `'static` is impossible, before D73077096.

Now we have `FunctionSignature`, it's actually possible to relax the `In`
lifetime. The code becomes more verbose, but it is nice to be able to pass
`&str`, not `String` to `call_function`

Reviewed By: zzl0

Differential Revision: D73078788

fbshipit-source-id: cb0e6e1a0a93ba6a1063b8be844a0db413ab9a41