git: don't clean up existing folder on "init"
Summary:
If "sl init --git" fails, we clean up the clone destination folder on error assuming it was an incomplete clone. However, you can "sl init --git" in an existing directory with contents, and if that operation fails we were incorrectly "cleaning up" the existing directory, which may have been files not under version control.

Fix logic to only clean up destination path if it was empty to start with.

Reviewed By: zzl0

Differential Revision: D72998541

fbshipit-source-id: dcd207b7053c4edf76f0b35d3892a4ab95f1925f