Add caching of status computation
Summary:
# Context:

We are working to refactor the EdenFs CLI client library into a supported XFN consumable library to avoid code duplication and sub-optimal usage patterns.

# This Diff:

Watchman had some caching of frequently ran Sapling operations. This will be handly for clients that are more long-running (e.g. buck2 and meerkat in daemon mode). Here we are adding caching to status results.

Reviewed By: lXXXw

Differential Revision: D73058808

fbshipit-source-id: bc11821a28cb64493e865a2d12a1ea39fbd7e55f