Add eslint rule to catch missing type annotations
Summary:
See previous diff for examples.

When we mirror to OSS, we make Internal act like `Record<string, never>` / empty object. So when we do `Internal.foo?.().then(param => ...)`, param will become `any`, even though it's really `never`. The  fix is to add an annotation, but it's easy to forget since it's a non-issue in internal builds where `Internal` is well-typed.

This also applies to `.catch`.

This eslint rule makes sure we don't miss these.

Reviewed By: nsblake

Differential Revision: D73122449

fbshipit-source-id: 7bb41f0823c24d36172f3e564b3e0d2a17689c7a