Move file system and database IO benchmarks here
Summary:
# Context

`eden debug bench` subcommand runs performance benchmarks for:
1. EdenFS
2. OS-native file system Eden is relying on Linux, macOS, and Windows.
3. Local databases Eden is using or is considering to use.

The benchmarks are intended to be cross-platform and should work on Linux, macOS, and Windows.

The benchmarks are intended to be highly reproducible, micro benchmarks, which we can use for performance testing both in CI and locally.

# This diff

Adds file system and database IO benchmarks.

Reviewed By: jdelliot

Differential Revision: D73298015

fbshipit-source-id: 1f991add08da761bdee4709afd67b5b4198dc54a