Add SQLite database IO benchmark
Summary:
# Context

`eden debug bench` subcommand runs performance benchmarks for:
1. EdenFS
2. OS-native file system Eden is relying on Linux, macOS, and Windows.
3. Local databases Eden is using or is considering to use.

The benchmarks are intended to be cross-platform and should work on Linux, macOS, and Windows.

The benchmarks are intended to be highly reproducible, micro benchmarks, which we can use for performance testing both in CI and locally.

# This diff

Add SQLite database IO benchmark

Reviewed By: jdelliot

Differential Revision: D73303086

fbshipit-source-id: 6355b207806c5c403b832905495006527cc8e354