Add a subcommand which for now does nothing.
Summary:
Similar to `eden debug stress` we are adding the subcommand for running benchmarks relevant to Eden, as well as the OS-native file systems Eden uses under the on Mac, Windows and Linux.

The benchmarks are intended to be cross-platform and should work on Windows, Mac and Linux.

The benchmarks are intended to be highly reproducible, micro benchmarks, which we can user for performance testing both in CI and local testing.

Reviewed By: jdelliot

Differential Revision: D73275161

fbshipit-source-id: 53a02a686604289edf97ece277dbcab2265f0c9b