Split FSIO and DBIO benchmarks into separate subcommands
Summary:
# Context

`eden debug bench` subcommand runs performance benchmarks for:
1. EdenFS
2. OS-native file system Eden is relying on Linux, macOS, and Windows.
3. Local databases Eden is using or is considering to use.

The benchmarks are intended to be cross-platform and should work on Linux, macOS, and Windows.

The benchmarks are intended to be highly reproducible, micro benchmarks, which we can use for performance testing both in CI and locally.

# This diff

Splits filesystem and database benchmarks into separate subcommands.

Reviewed By: jdelliot

Differential Revision: D73299556

fbshipit-source-id: 02e082b1a5790e23c78b99754365dfd05627b66d