tests: ignore fbthrift related errors
Summary:
Example error on Windows on D73210181's `sapling-oss-windows` test:

  error[E0046]: not all trait items implemented, missing: `rs_thrift_write`
      --> fbcode\eden\scm\build\cargo-target\debug\build\scope-ea80f1dfd87ff135\out\lib.rs:1550:1
       |
  1550 | / impl<P> ::fbthrift::Serialize<P> for self::Structured
  1551 | | where
  1552 | |     P: ::fbthrift::ProtocolWriter,
       | |__________________________________^ missing `rs_thrift_write` in implementation
       |
       = help: implement the missing item: `fn rs_thrift_write(&self, _: &mut P) { todo!() }`

It seems to be a mismatch of some kind of library version or something.
It's not very interesting (not in our code). So let's just ignore it for now.

The output can be very long (2k lines in this case). So I added a way to truncate it.

Reviewed By: muirdm

Differential Revision: D73388514

fbshipit-source-id: b4563d8c6616cb3adec581fff1861d6cb117a7fb