Remove summary printing as individual benchmark displays contain the information needed
Summary:
# Context

`eden debug bench` subcommand runs performance benchmarks for:
1. EdenFS
2. OS-native file system Eden is relying on Linux, macOS, and Windows.
3. Local databases Eden is using or is considering to use.

The benchmarks are intended to be cross-platform and should work on Linux, macOS, and Windows.

The benchmarks are intended to be highly reproducible, micro benchmarks, which we can use for performance testing both in CI and locally.

# This diff

Remove summary printing as individual benchmark displays contain the information needed

Reviewed By: jdelliot

Differential Revision: D73395981

fbshipit-source-id: 79ff8cb48fb8be6e546f5286622acfac31422a9a