bench.rs is getting large, refactor into separate modules
Summary:
# Context

`eden debug bench` subcommand runs performance benchmarks for:
1. EdenFS
2. OS-native file system Eden is relying on Linux, macOS, and Windows.
3. Local databases Eden is using or is considering to use.

The benchmarks are intended to be cross-platform and should work on Linux, macOS, and Windows.

The benchmarks are intended to be highly reproducible, micro benchmarks, which we can use for performance testing both in CI and locally.

# This diff

bench.rs is getting large, refactor into separate modules

Reviewed By: jdelliot

Differential Revision: D73403105

fbshipit-source-id: 96ff2ccbeadb0292b1768ed343b85e8107074fa7