Introduce Enum for benchmark types
Summary:
# Context

`eden debug bench` subcommand runs performance benchmarks for:
1. EdenFS
2. OS-native file system Eden is relying on Linux, macOS, and Windows.
3. Local databases Eden is using or is considering to use.

The benchmarks are intended to be cross-platform and should work on Linux, macOS, and Windows.

The benchmarks are intended to be highly reproducible, micro benchmarks, which we can use for performance testing both in CI and locally.

# This diff

Replace String with Enum, as suggested by jdelliot

Reviewed By: jdelliot

Differential Revision: D73393144

fbshipit-source-id: 0e531719dee98b9f8ab97678a2790712e9c2e9a1