First version of the traversal benchmark
Summary:
# Context

`eden debug bench` subcommand runs performance benchmarks for:
1. EdenFS
2. OS-native file system Eden is relying on Linux, macOS, and Windows.
3. Local databases Eden is using or is considering to use.

The benchmarks are intended to be cross-platform and should work on Linux, macOS, and Windows.

The benchmarks are intended to be highly reproducible, micro benchmarks, which we can use for performance testing both in CI and locally.

# This diff

First version of the traversal benchmark

Reviewed By: jdelliot

Differential Revision: D73405780

fbshipit-source-id: 9cf6771d535005911667670396f6960c51fa1d6b