In RocksDB benchmark remove options object creation inside the measured code
Summary:
# Context

`eden debug bench` subcommand runs performance benchmarks for:
1. EdenFS
2. OS-native file system Eden is relying on Linux, macOS, and Windows.
3. Local databases Eden is using or is considering to use.

The benchmarks are intended to be cross-platform and should work on Linux, macOS, and Windows.

The benchmarks are intended to be highly reproducible, micro benchmarks, which we can use for performance testing both in CI and locally.

# This diff

In RocksDB benchmark remove options object creation inside the measured code

Reviewed By: jdelliot

Differential Revision: D73320777

fbshipit-source-id: 17969fb431ad74f8d20a2e8348a0f71385af1f97