Move the display function into a Display trait
Summary:
# Context

`eden debug bench` subcommand runs performance benchmarks for:
1. EdenFS
2. OS-native file system Eden is relying on Linux, macOS, and Windows.
3. Local databases Eden is using or is considering to use.

The benchmarks are intended to be cross-platform and should work on Linux, macOS, and Windows.

The benchmarks are intended to be highly reproducible, micro benchmarks, which we can use for performance testing both in CI and locally.

# This diff

Move the display function into a Display trait

Reviewed By: jdelliot

Differential Revision: D73394632

fbshipit-source-id: ee5011663ec22d24533529b4dfe0c5fac21ed7ee