Refactor TestDir into a separate data type
Summary:
# Context

`eden debug bench` subcommand runs performance benchmarks for:
1. EdenFS
2. OS-native file system Eden is relying on Linux, macOS, and Windows.
3. Local databases Eden is using or is considering to use.

The benchmarks are intended to be cross-platform and should work on Linux, macOS, and Windows.

The benchmarks are intended to be highly reproducible, micro benchmarks, which we can use for performance testing both in CI and locally.

# This diff

Starting the refactoring of the benchmark into a library.

Start with separating RandomData and TestDir into separate data types.

Reviewed By: jdelliot

Differential Revision: D73325840

fbshipit-source-id: 5103d77aef39c94a169a6d51872cab8a7472fada