Seperate benchmark running and displaying
Summary:
# Context

`eden debug bench` subcommand runs performance benchmarks for:
1. EdenFS
2. OS-native file system Eden is relying on Linux, macOS, and Windows.
3. Local databases Eden is using or is considering to use.

The benchmarks are intended to be cross-platform and should work on Linux, macOS, and Windows.

The benchmarks are intended to be highly reproducible, micro benchmarks, which we can use for performance testing both in CI and locally.

# This diff

Making progress towards turning the monolithic binary into a library of benchmarks. As we will be adding more benchmarks and the `bench.rs` is getting large.

Separate benchmark running and displaying.

Reviewed By: jdelliot

Differential Revision: D73370690

fbshipit-source-id: 7e6f433d01df3bcde9049171d969f2c741153c2f