display Benchmark with nice alignment
Summary:
# Context

`eden debug bench` subcommand runs performance benchmarks for:
1. EdenFS
2. OS-native file system Eden is relying on Linux, macOS, and Windows.
3. Local databases Eden is using or is considering to use.

The benchmarks are intended to be cross-platform and should work on Linux, macOS, and Windows.

The benchmarks are intended to be highly reproducible, micro benchmarks, which we can use for performance testing both in CI and locally.

# This diff

display Benchmark with nice alignment

Reviewed By: jdelliot

Differential Revision: D73373703

fbshipit-source-id: 788ff8560364f989f279d00514c2d900960a1079