Refactoring by renaming measurement to metric for consistency
Summary:
# Context

`eden debug bench` subcommand runs performance benchmarks for:
1. EdenFS
2. OS-native file system Eden is relying on Linux, macOS, and Windows.
3. Local databases Eden is using or is considering to use.

The benchmarks are intended to be cross-platform and should work on Linux, macOS, and Windows.

The benchmarks are intended to be highly reproducible, micro benchmarks, which we can use for performance testing both in CI and locally.

# This diff

Refactoring by renaming measurement to metric for consistency

Reviewed By: jdelliot

Differential Revision: D73415916

fbshipit-source-id: 8ca2cf5b43d813904d557d31a97115d591a62f39