Add open() and read() benchmark to the traversal benchmark
Summary:
# Context

`eden debug bench` subcommand runs performance benchmarks for:
1. EdenFS
2. OS-native file system Eden is relying on Linux, macOS, and Windows.
3. Local databases Eden is using or is considering to use.

The benchmarks are intended to be cross-platform and should work on Linux, macOS, and Windows.

The benchmarks are intended to be highly reproducible, micro benchmarks, which we can use for performance testing both in CI and locally.

# This diff

Add open() and read() benchmark to the traversal benchmark

Reviewed By: jdelliot

Differential Revision: D73419583

fbshipit-source-id: e93afcc430f6632b0bc55da2395a4f929d31c4ba