In the traversal benchmark update the progress bar every 100 files
Summary:
# Context

`eden debug bench` subcommand runs performance benchmarks for:
1. EdenFS
2. OS-native file system Eden is relying on Linux, macOS, and Windows.
3. Local databases Eden is using or is considering to use.

The benchmarks are intended to be cross-platform and should work on Linux, macOS, and Windows.

The benchmarks are intended to be highly reproducible, micro benchmarks, which we can use for performance testing both in CI and locally.

# This diff

In the traversal benchmark update the progress bar every 100 files

Reviewed By: jdelliot

Differential Revision: D73417264

fbshipit-source-id: f77b09cf28ae9afa0676ac682e6ded13dc13ed94