Use pub-sub model for updating state in WBC
Summary:
With all the work done so far, we can finally implement pub-sub style WBC update as an alernative for direct polling model. This has a great potential of significantly reducing the DB load generated by the consistent polling of Mononoke XDB by thousands of Git repos, especially for Mononoke Git and SCS server

Note that integration tests validating this behaviour are not possible right now, because we have only two modes: Prod and  LogToFile both of which are insufficent for our use-case. I will make the necessary changes in follow up diffs to include a separate `InMemory` mode which should help with this.

Reviewed By: YousefSalama

Differential Revision: D73109741

fbshipit-source-id: daa144cc945eccbeb077a4d5a44092e2f231b329