Enable scribe backed caching for BonsaiTagMapping
Summary:
We have all the pieces in place, let's start using `CachedBonsaiTagMapping` instead of the direct SQL one to avoid making so many tag fetch calls to the DB.
Note that this functionality is behind a JK so safe to land. It will only be enabled in production once the JK turns on.

Reviewed By: YousefSalama

Differential Revision: D73449817

fbshipit-source-id: 124c9b94535596e190a53465d17cf110fa8f438f