Introduce NotificationFilter for RepoEventPublisher
Summary:
It is possible that multiple metadata caches might want to get updated based on the same scribe category, however the current design does not support that mode. It implicitly assumes that every new variant of the metadata cache will have a different scribe category to listen to.

This diff introduces the notion of `NotificationFilter` which allows the user to determine which types of payload the subscriber wants to get notified on within the same category. I am going to make use of this to create a pub-sub based cached version of BonsaiTagMapping with the same category that's used for WBC

Reviewed By: YousefSalama

Differential Revision: D73439600

fbshipit-source-id: 7df6f2060f809b125ec504f1c45c2fb5bc89dfbd