Factory construction of Repo Event Publisher facet
Summary: This diff provides the basic factory construction of Repo Event Publisher facet. Note that in the current implementation, there is duplication, i.e. for the same configuration we would construct a new instance of this facet per individual repo which would be incredibly wasteful. Follow up diff will fix that.

Reviewed By: YousefSalama

Differential Revision: D73050768

fbshipit-source-id: 31ac6859f1657ed87feb96dd59c4bfec17938e47