Include Repo Event Publisher as part of Repo facet container and WBC repo facet
Summary:
This diff just makes `RepoEventPublisher` a part of the main Repo facet container used by Mononoke Server, SCS and Mononoke Git service. It also adds this facet as a dependency for constructing the `WarmBookmarksCache` facet. This itself was sufficient boilerplate that I am confining it to its own diff.

In follow up diff, I will start leveraging `RepoEventPublisher` in WBC to make it more efficient.

Reviewed By: YousefSalama

Differential Revision: D73105767

fbshipit-source-id: 19850380150367529ce2107f48c5091a6971b2a3