add a config for leaking outstanding thrift requests
Summary:
# This diff

Uses a config to determine whether we should enable a risky Thrift config that allows the EdenFS thrift server to leak requests after the Eden thrift server has been stopped.

# Context

We believe the root cause of a very common crash (~900 crashes per month on macOS, ~70k per month on Linux) stems from the use of the `leakOutstandingRequestsWhenServerStops` thrift config.

The crash is a memory safety issue caused by attempting to dereference a Thrift Server member variable after the server has been deconstructed. The member variable is dereferenced when outstanding requests are processed. If we remove any processing of outstanding requests after the server is stopped, then it should be impossible to hit this crash.

Reviewed By: jdelliot, kavehahmadi60

Differential Revision: D73440987

fbshipit-source-id: 04bc3cdd11a0d1e606dac03cc41e52f969f82057