add getProtocolCapabilitiesWithoutTrim
Summary:
## Context:
EdenFS crash during graceful restart when the `TakeoveData` is large. One solution for this crash is to spilit the `TakeoverData` into 1 GB chunks, and send data over socket in chunks. More explanations and pseudocode is in this [post](https://fb.workplace.com/groups/1708850869939124/permalink/1952362862254589/)

## This diff:
The `getProtocolCapabilities` read the header of the message and return the protocol capability but it trims the data from the beginning of the message to prepare the real data for deserliazation. In the next diffs, we need a function that peak the header and return the protocol capability, while it doesn't trim the data and keep it. It will be used when we want to read the protocol capability of the message multiple time.
This diff adds a new function in `TakeoverData` to get the protocol capability while it doesn't trim the data

Reviewed By: jdelliot

Differential Revision: D73276683

fbshipit-source-id: 85910dac11fd3968078014408c416a13d4000b76