add isChunked and CHUNKED_MESSAGE capability
Summary:
## Context:
EdenFS crash during graceful restart when the `TakeoveData` is large. One solution for this crash is to spilit the `TakeoverData` into 1 GB chunks, and send data over socket in chunks. More explanations and pseudocode is in this [post](https://fb.workplace.com/groups/1708850869939124/permalink/1952362862254589/)

## This diff:
This diff add two things:
1- Add a new capability `TakeoverCapabilities::CHUNKED_MESSAGE`. For now this capability is not part of any version, and it is not in production.
2- Adds `TakeoverData::isChunked`. This bool function is used to check if the server send the data in chunks. For now, no server send the data in chunks. Therefore this function return false for all the messages.

Reviewed By: jdelliot

Differential Revision: D73279304

fbshipit-source-id: bb0fad8069c3001e8669528c4efb57d3d2beca6a